separate folder for controllers.js
